requirements.txt
asgiref==3.8.1
asn1crypto==1.5.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.1
Django==5.1.2
djangorestframework==3.15.2
filelock==3.16.1
idna==3.10
packaging==24.1
platformdirs==4.3.6
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pytz==2024.2
requests==2.32.3
snowflake-connector-python==3.12.2
sortedcontainers==2.4.0
sqlparse==0.5.1
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.2.3
